Speed up IsCheck by "ray-tracing" from the king
The previous implementation went through all the possible moves for the
opposing colour and computed whether it was king capture or not.
That is less efficient than "ray-tracing" from the king when most of the
pieces are still on the board, ie. during the opening. An optimal
IsCheck algorithm would possibly take into account which pieces are
still on the board, but this implementation is already a lot faster for
the opening stage of the game (and the middle/late game is faster anyway
because there are less pieces on the board).